Simplification of cyclops library
- Also found that I had the OC and VI pins swiched on the Teensy pinout
(the OC pins were tied to A10-13 which are _only_ capable of analog
input). I simply swiched the pins and reflected that change in the
library. Now A10-13 are used for VI which is only analog in.
- For now, I will leave the GSOC dev stuff on its own branch and remove
it from the repo for the 3.6 production. Hopefully it can be integrated
in the future. It is too complicated and unfinished to be merged with
master right now.